requirements.txt
requests
python-dotenv
sendgrid
beautifulsoup4